ingestion/pyproject.toml
[project]
name = "ingestion"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "google-cloud-bigquery>=3.41.0",
    "google-cloud-storage>=3.10.1",
    "pandas>=3.0.2",
    "pyarrow>=23.0.1",
    "requests>=2.33.1",
]
